refactor users changing  methods - after review
