+++ src/common/commands/block-user.command.ts
@@ -1,0 +1,3 @@
+export class BlockUserCommand {
+  constructor(public readonly userId: string) {}
+}
