+++ src/common/commands/deactivate-admin.command.ts
@@ -1,0 +1,3 @@
+export class DeactivateAdminCommand {
+  constructor(public readonly userId: string) {}
+}
